pyproject.toml
[tool.poetry]
name = "cncrd-ext-stats"
version = "3.0.0"
description = "Statistics extension for Concord"
authors = ["Nariman Safiulin <woofilee@gmail.com>"]
license = "MIT"

readme = "README.md"
repository = "https://github.com/Roolat/concord-ext-stats"

keywords = ["discord", "concord", "bot", "stats"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: Russian",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [
    { include = "concord" },
]

[tool.poetry.dependencies]
python = "^3.6"
cncrd = "^1.0.0"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", branch = "rewrite" }
pendulum = "^2.0"

[tool.poetry.dev-dependencies]
black = "^18.9b0"
pre-commit = "^1.10"

[tool.black]
line-length = 80
py36 = true
